Removed MaybeOwningRef's copy constructor
While the copy constructor was well-defined and had intuitive semantics,
it was prone to causing bugs because its presence could cause you to
forget to move a MaybeOwningRef, and if it was owning, when the middle
one goes out of scope (which usually would happen if you meant to move
from it), the resulting reference would be dangling.  The behavior of
the copy constructor is still easy to get by using
MaybeOwningRef(*other), so use that if you want to copy a
MaybeOwningRef.