Big lifetime refactor
Instead of objects being owned by the user of the library, objects are
now shared between whatever wants them.  Notable possible owners are the
user, the scene, and animations.  Objects will be automatically deleted
by the scene if it finds that it is the only thing left owning it.
Animations now always add their objects and will never remove them,
meaning that objects no longer have to be explicitly added to the scene
before use.  Animations don't have to bother with removing their objects
since the scene will do that automatically for any objects that the
animation had made.

This has all been done with a new shared_ptr-like class called
ObjectPtr.  Its semantics are pretty similar to shared_ptr, although
there's a couple differences.  All objects should get made with helper
functions now that wrap the constructor arguments.

I'm still running into a bug, but I've noticed that it doesn't get
caught by any of the current tests and I was seeing similar things
before this refactor, so I'm thinking the bug is unrelated to the
refactor and I'll try to fix it in the next commit.