Fixed z buffer in texture_transform
I really need to find a better way to keep the vertex calculations
global for these exotic shaders because this is the second time I've
fixed a bug like this.

This fixed some of the issues I mentioned in my last commit message, but
some of the issues still haven't been fixed.