Updated CMakeLists with a different structure
As I've developed ganim more, my ideas for how exactly to implement it
have changed, so I needed to redo the project structure considerably.
There are now three main components: libganim, ganimscript, and ganim.
libganim is the base animation engine, ganimscript will be a scripting
language that is a wrapper of sorts around libganim, and then ganim is
the graphical editor that uses ganimscript to interface with libganim.

In CMakeLists.txt, things are confusing to allow for tests.  ganimscript
and ganim are split into a library and an executable to allow for the
test executable to link to the contents of ganimscript and ganim without
getting their mains.