Added definitions of unicode identifier characters
They're mostly based on Unicode standard annex 31's definition of
XID_Start and XID_Continue.  I'm not 100% sure I implemented it
correctly, but for now it should be good.  One interesting thing is that
for some reason, ·, which I've always used for the dot product, is
considered an Other_ID_Continue, when I want it to be considered an
operator, so I removed that from XID_Continue.