Re-License to full GPL
Initially I thought of ganim as a library, and even the future editor was going to have a library-like component that people could link to to extend it.  I generally prefer the LGPL over the GPL for libraries because I don't like forcing others to use a particular license for code that is not derived from my own.  However, I've realized that people could make "plugins" for the editor by just writing ganimscript, not C++, so that part of the argument is largely moot.  I thought of licensing different parts of the code differently, so that the underlying library was licensed with the LGPL and the rest was licensed with the GPL, but that seemed confusing for both me and others so I decided I would just go with the GPL for everything.