Changed the way GeX spacing works slightly
It appears that TeX adds kerning even for symbols that are not right
next to each other.  This seems like it might be a bug, but that's how
it is.  Since I'm using harfbuzz, this idea doesn't really apply to me,
so I can't copy this behavior.  I noticed the discrepancy earlier and
multiplied my em width by 1.2 to try to compensate, but that isn't a
good solution for other measurements.  Instead, the em width is not
getting multiplied by anything, and whenever I'm not using harfbuzz to
do spacing, I am just adding one mu of space to whatever the spacing
should be.  This isn't perfect, but it should be good enough.